Struct compute1::Operation

pub struct Operation {
    pub status: Option<String>,
    pub insert_time: Option<String>,
    pub warnings: Vec<OperationWarnings>,
    pub error: Option<OperationError>,
    pub target_id: Option<String>,
    pub target_link: Option<String>,
    pub start_time: Option<String>,
    pub client_operation_id: Option<String>,
    pub creation_timestamp: Option<String>,
    pub id: Option<String>,
    pub kind: Option<String>,
    pub name: Option<String>,
    pub zone: Option<String>,
    pub region: Option<String>,
    pub self_link: Option<String>,
    pub operation_type: Option<String>,
    pub http_error_message: Option<String>,
    pub progress: Option<i32>,
    pub end_time: Option<String>,
    pub http_error_status_code: Option<i32>,
    pub status_message: Option<String>,
    pub user: Option<String>,
}

An operation resource, used to manage asynchronous API requests.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

delete target instances (response)

stop instances (response)

add instance target pools (response)

add health check target pools (response)

remove health check target pools (response)

update backend services (response)

delete target pools (response)

update firewalls (response)

insert networks (response)

insert addresses (response)

patch backend services (response)

set tags instances (response)

set disk auto delete instances (response)

insert target http proxies (response)

set usage export bucket projects (response)

update url maps (response)

insert instance templates (response)

patch firewalls (response)

delete firewalls (response)

detach disk instances (response)

delete access config instances (response)

delete backend services (response)

insert instances (response)

set target global forwarding rules (response)

insert url maps (response)

add access config instances (response)

get region operations (response)

insert backend services (response)

insert disks (response)

delete instance templates (response)

delete global forwarding rules (response)

insert global forwarding rules (response)

delete disks (response)

delete global addresses (response)

set common instance metadata projects (response)

insert target pools (response)

delete target http proxies (response)

delete networks (response)

get global operations (response)

delete url maps (response)

set scheduling instances (response)

attach disk instances (response)

delete addresses (response)

reset instances (response)

set url map target http proxies (response)

create snapshot disks (response)

insert routes (response)

insert global addresses (response)

deprecate images (response)

patch http health checks (response)

delete http health checks (response)

set metadata instances (response)

insert http health checks (response)

delete images (response)

delete snapshots (response)

insert forwarding rules (response)

insert images (response)

start instances (response)

set backup target pools (response)

insert target instances (response)

delete forwarding rules (response)

move instance projects (response)

set target forwarding rules (response)

move disk projects (response)

insert firewalls (response)

remove instance target pools (response)

delete instances (response)

get zone operations (response)

update http health checks (response)

delete routes (response)

patch url maps (response)

Fields

status

[Output Only] Status of the operation. Can be one of the following: PENDING, RUNNING, or DONE.

insert_time

[Output Only] The time that this operation was requested. This is in RFC3339 text format.

warnings

[Output Only] If warning messages are generated during processing of the operation, this field will be populated.

error

[Output Only] If errors are generated during processing of the operation, this field will be populated.

target_id

[Output Only] Unique target ID which identifies a particular incarnation of the target.

target_link

[Output Only] URL of the resource the operation is mutating.

start_time

[Output Only] The time that this operation was started by the server. This is in RFC3339 text format.

client_operation_id

[Output Only] An optional identifier specified by the client when the mutation was initiated. Must be unique for all operation resources in the project

creation_timestamp

[Output Only] Creation timestamp in RFC3339 text format.

id

[Output Only] Unique identifier for the resource; defined by the server.

kind

[Output Only] Type of the resource. Always compute#Operation for Operation resources.

name

[Output Only] Name of the resource.

zone

[Output Only] URL of the zone where the operation resides.

region

[Output Only] URL of the region where the operation resides. Only applicable for regional resources.

self_link

[Output Only] Server defined URL for the resource.

operation_type

[Output Only] Type of the operation, such as insert, update, and delete.

http_error_message

[Output Only] If the operation fails, this field contains the HTTP error message that was returned, such as NOT FOUND.

progress

[Output Only] An optional progress indicator that ranges from 0 to 100. There is no requirement that this be linear or support any granularity of operations. This should not be used to guess at when the operation will be complete. This number should be monotonically increasing as the operation progresses.

end_time

[Output Only] The time that this operation was completed. This is in RFC3339 text format.

http_error_status_code

[Output Only] If the operation fails, this field contains the HTTP error message that was returned, such as 404.

status_message

[Output Only] An optional textual description of the current status of the operation.

user

[Output Only] User who requested the operation, for example: user@example.com.