Struct compute1::ProjectMethodsBuilder

pub struct ProjectMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the Compute hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate compute1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use compute1::Compute;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Compute::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `get(...)`, `move_disk(...)`, `move_instance(...)`, `set_common_instance_metadata(...)` and `set_usage_export_bucket(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, NC, A> ProjectMethodsBuilder<'a, C, NC, A>

fn move_disk(&self, request: &DiskMoveRequest, project: &str) -> ProjectMoveDiskMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Moves a persistent disk from one zone to another.

fn set_common_instance_metadata(&self, request: &Metadata, project: &str) -> ProjectSetCommonInstanceMetadataMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Sets metadata common to all instances within the specified project using the data included in the request.

fn get(&self, project: &str) -> ProjectGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns the specified project resource.

fn move_instance(&self, request: &InstanceMoveRequest, project: &str) -> ProjectMoveInstanceMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Moves an instance and its attached persistent disks from one zone to another.

fn set_usage_export_bucket(&self, request: &UsageExportLocation, project: &str) -> ProjectSetUsageExportBucketMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Enables the usage export feature and sets the usage export bucket where reports are stored. If you provide an empty request body using this method, the usage export feature will be disabled.